requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
greenlet==3.0.3
grpcio==1.65.4
grpcio-tools==1.65.4
idna==3.7
overrides==7.7.0
playwright==1.46.0
protobuf==5.27.3
pyee==11.1.0
PyYAML==6.0.2
requests==2.32.3
robotframework==7.0.1
robotframework-assertion-engine==3.0.3
robotframework-browser==18.7.0
robotframework-pythonlibcore==4.4.1
setuptools==72.2.0
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
